Rename ui files to match simpler scheme
