+++ packages/components/src/ui/flex/use-flex-item.js
@@ -7,7 +7,7 @@ import { css, cx, ui } from '@wp-g2/styles';
 /**
  * Internal dependencies
  */
-import * as styles from './flex-styles';
+import * as styles from './styles';
 
 /**
  * @param {import('@wp-g2/create-styles').ViewOwnProps<import('./types').FlexItemProps, 'div'>} props
