+++ packages/components/src/ui/grid/index.js
@@ -1,2 +1,2 @@
-export { default as Grid } from './grid';
-export { default as useGrid } from './use-grid';
+export { default as Grid } from './component';
+export { default as useGrid } from './hook';
