+++ packages/components/src/ui/h-stack/hook.js
@@ -9,7 +9,7 @@ import { getValidChildren } from '@wp-g2/utils';
  * Internal dependencies
  */
 import { FlexItem, useFlex } from '../flex';
-import { getAlignmentProps } from './h-stack-utils';
+import { getAlignmentProps } from './utils';
 
 /**
  *
