+++ packages/components/src/ui/text/component.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { createComponent } from '../utils';
-import useText from './use-text';
+import useText from './hook';
 
 const Text = createComponent( {
 	as: 'span',
