+++ packages/components/src/ui/text/index.js
@@ -1,2 +1,2 @@
-export { default as Text } from './text';
-export { default as useText } from './use-text';
+export { default as Text } from './component';
+export { default as useText } from './hook';
