+++ packages/components/src/ui/text/types.ts
@@ -1,4 +1,4 @@
-import type { Props as TruncateProps } from '../truncate/types';
+import type { Props as TruncateProps } from '../truncate/hook';
 import type { CSSProperties } from 'react';
 
 type TextAdjustLineHeightForInnerControls =
