+++ packages/components/src/ui/truncate/hook.js
