+++ packages/components/src/ui/truncate/index.js
@@ -1,2 +1,2 @@
-export { default as Truncate } from './truncate';
-export { default as useTruncate } from './use-truncate';
+export { default as Truncate } from './component';
+export { default as useTruncate } from './hook';
