+++ packages/components/src/ui/truncate/test/truncate.js
@@ -6,7 +6,7 @@ import { render } from '@testing-library/react';
 /**
  * Internal dependencies
  */
-import Truncate from '../truncate';
+import { Truncate } from '..';
 
 describe( 'props', () => {
 	test( 'should render correctly', () => {
