+++ packages/components/src/ui/v-stack/index.js
@@ -1,3 +1,3 @@
-export { default as VStack } from './v-stack';
+export { default as VStack } from './component';
 
-export * from './use-v-stack';
+export * from './hook';
